Fix load module for tests
